Moved shared code to tidb_query_shared_expr.
