+++ components/tidb_query_vec_expr/Cargo.toml
@@ -16,7 +16,6 @@ num = { version = "0.2", default-features = false }
 num-traits = "0.2"
 openssl = { version = "0.10" }
 protobuf = "2"
-safemem = { version = "0.3", default-features = false }
 tidb_query_codegen = { path = "../tidb_query_codegen" }
 tidb_query_datatype = { path = "../tidb_query_datatype" }
 tidb_query_common = { path = "../tidb_query_common" }
